Ensure figcaptions aren’t hidden by figures
A crude way of ensuring that figcaptions are never covered by images/videos with a slightly-off padding-top value that’s necessary for lazy loading. At the moment it relies on b-lazy’s addition of b-error to the main figure element, which means the figure has been loaded and the padding-top value can be overriden. Not the best way of doing it, but works for now.