Fixed images not uploading
This took a while to figure out, but was a simple fix.

The savePage function now starts after ContentTools is finished converting its weird div-based HTML to regular HTML with <img> tags and whatnot. Calling savePage before that happened was causing weird issues, in part because to-markdown ignores <div>’s that are empty, which is what ContentTools does when manipulating images.

So now it works properly. Images are detected by to-markdown and coverted to Kirby syntax successfully.

Still need to add the other JS files that make editing work to the repo.