Fixed showing empty widgets to page editors
The groups widget used to be visible to editors of a page even when no groups were associated with that page.

Now, empty widgets are hidden by default with a “hidden” class. If the editor adds something to that widget to make it not empty anymore, the “hidden” class is removed so that it’s still visible when the user clicks the save button.