Ability to “join” and administer group pages
Users who are logged in but not part of a group/class can now request access. Their request is stored in the page’s UserData field, and shows up to existing authors when they refresh the page while logged in.

Existing authors can approve new authors or deny their requests.