New feature: file uploads
Added an additional tool to the inline editing toolbox; files.

Clicking the icon triggers a click on the upload form’s input. When a file is selected, AJAX is used to push the file to a new “uploadnew” route within config.php, which places the file in the right place and returns back some JSON with the file’s location and filename.

The returning JSON triggers a new element to be created via ContentTools so that it can be manipulated. The textual description of the file can be changed just by tweaking the name that appears.

This doesn’t support uploading multiple items at once, however.