New features: Font settings and mobile TOC
Font settings (top-right corner of some pages) allow the user to toggle dark mode, switch to a dyslexic font, and increase the font size. These settings are stored within the user’s web browser and carry over to all other pages. Eventually when PJAX page switching is implemented the transition between pages should be seamless.

Also added a button to the top-left corner that allows mobile users to access the page’s Table of Contents for quick navigation to headings. This currently has some bugs, but will eventually appear only on mobile and will be dismissable by tapping outside of the region or by scrolling down.