Improved authors widget search
Brought the author search/add experience in line with the recent Group and Event widgets. Also changed terminology from “authors” to “users”, which is a more context agnostic term.

Should improve the code within main.js in particular. Right now it’s a bit difficult to follow, but it work.