New: Automatically add related pages to new page
When a new project, event, handbook, or anything else is created from within another user-created page (like a group page), the newly-created page will automatically be “related” to the other page and tied together.

Basically, this allows projects created from within a group’s page to automatically display that group’s name in the sidebar of the new project. This eliminates the need to manually tag projects as being part of a group or event or whatever.