Major rework of site-wide navigation
Instead of hard-coding every menu item and using a bunch of logic to determine when each one should be “active”, I’ve greatly simplified the logic and turned the menu into an array within network-methods.php for now. It’s still a bit complex, but it’s better, and far easier to manipulate pages.

This is the first step in making the navigation area entirely editable by the site admin. Eventually they should be able to click and drag menu items around, set titles, and set manual URLs all from within the menu area.

I also changed the HTML of the navigation to position sub-items within the <li> of main items. This is the more accessible and screen-readable way of setting up the nav, but I’ve noticed some styling problems on mobile and in small windows on the desktop. I may need to investigate using ARIA to tie two disparate DIVs together to achieve the layout I want while still being accessible.